requirements.txt
# Core utilities
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0

# LLMs and RAG
google-generativeai>=0.8.0
langchain>=0.3.0
chromadb>=1.1.0
sentence-transformers>=2.2.0

# Visualization
matplotlib>=3.7.0

# Web Interface
streamlit>=1.50.0

# Development and Testing
pytest>=7.4.0
pytest-mock>=3.11.0
